+++ CHANGELOG.md
@@ -1,10 +1,16 @@
+## 0.5.0
+
+* New ddToDMS current instance converter 
+* New dmsToDD current instance converter
+* Updates example
+
 ## 0.3.0
 
 * Creates ddToDMS feature 
 * Creates dmsToDD feature
 * Creates ddToUTM feature
-* Update README
-* Add example
+* Updates README
+* Adds example
 
 ## 0.0.1